+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [60.1.0]
 - [Bottomsheet] Added page navigation inside bottomsheet
+- [Checkmark] Added `AddStrokeToCheckmark` property to display a visible stroke around the checkmark
+- [CheckmarkListItem] Added `AddStrokeToCheckmark` property to display a visible stroke around the checkmark icon
 
 ## [60.1.0]
 - [DateAndTimePicker] `MaximumDate` and `MinimumDate` now also constrain the time portion when the selected date equals the boundary date.